requirements.txt
streamlit
newspaper3k
transformers
torch
nltk
pandas
scikit-learn
# Optional for explanation
shap
